Refactor payment:add-provider command
